% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_elastic_net.R
\name{calculate_lambda_max}
\alias{calculate_lambda_max}
\title{calculate data-based lambda.max}
\usage{
calculate_lambda_max(y, X, alpha)
}
\arguments{
\item{y}{the responding variable. The function assume it is non-missing}

\item{X}{the full feature dataset}

\item{alpha}{the alpha hyperparameter}
}
\value{
\item{name a}{description}
}
\description{
DESCRIPTION
}
\note{
The function calculate the minimal lambda that shrinks all
            regression coefficients into zero (lambda.max) based on the
            data:
}
