% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_elastic_net.R
\name{calculate_search_grid_en}
\alias{calculate_search_grid_en}
\title{calculate the whole grid for lasso}
\usage{
calculate_search_grid_en(y, X, n_alpha, n_lambda, lambda_ratio = NULL)
}
\arguments{
\item{y}{the responding variable. The function assume it is non-missing}

\item{X}{the full feature dataset}

\item{n_alpha}{number of alpha tested in the grid}

\item{n_lambda}{number of lambda tested in the grid}

\item{lambda_ratio}{Smallest value for lambda, as a fraction of lambda.max,}
}
\value{
a data frame
}
\description{
DESCRIPTION
}
\note{
A search grid calculated based on the input data
}
