% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_MLP.R
\name{calculate_search_grid_mlpL2}
\alias{calculate_search_grid_mlpL2}
\title{calculate the whole grid for MLP with L2 penalty}
\usage{
calculate_search_grid_mlpL2(y, X, n_size, n_lambda)
}
\arguments{
\item{y}{the responding variable. The function assume it is non-missing}

\item{X}{the full feature dataset}

\item{n_size}{number of size in grid search,}

\item{n_lambda}{number of lambda in grid search,}
}
\value{
a data frame
}
\description{
DESCRIPTION
}
\note{
A search grid calculated based on the input data
}
