% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_random_forest.R
\name{calculate_search_grid_rf}
\alias{calculate_search_grid_rf}
\title{calculate the whole grid for rf}
\usage{
calculate_search_grid_rf(y, X, n_mtry, n_MNS)
}
\arguments{
\item{y}{the responding variable. The function assume it is non-missing}

\item{X}{the full feature dataset}

\item{n_mtry}{number of mtry tested in the grid}

\item{n_MNS}{number of min.node.size tested in the grid}
}
\value{
a data frame
}
\description{
DESCRIPTION
}
\note{
A search grid calculated based on the input data
}
