% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_SVR.R
\name{calculate_search_grid_svrl}
\alias{calculate_search_grid_svrl}
\title{calculate the whole grid for SVR linear kernel}
\usage{
calculate_search_grid_svrl(y, X, n_C)
}
\arguments{
\item{y}{the responding variable. The function assume it is non-missing}

\item{X}{the full feature dataset}

\item{n_C}{number of Cost in grid search,}
}
\value{
a data frame
}
\description{
DESCRIPTION
}
\note{
A search grid calculated based on the input data
}
