% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_SVR.R
\name{calculate_search_grid_svrp}
\alias{calculate_search_grid_svrp}
\title{calculate the whole grid for SVR Polynomial kernel}
\usage{
calculate_search_grid_svrp(y, X, n_C, n_degree, n_Scale)
}
\arguments{
\item{y}{the responding variable. The function assume it is non-missing}

\item{X}{the full feature dataset}

\item{n_C}{number of Cost in grid search,}

\item{n_degree}{number of degree in grid search,}

\item{n_Scale}{number of Scale in grid search,}
}
\value{
a data frame
}
\description{
DESCRIPTION
}
\note{
A search grid calculated based on the input data
}
