% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_xgblinear.R
\name{calculate_search_grid_xgbl}
\alias{calculate_search_grid_xgbl}
\title{calculate the whole grid for xgblinear}
\usage{
calculate_search_grid_xgbl(y, X, n_rounds, n_reg)
}
\arguments{
\item{y}{the responding variable. The function assume it is non-missing}

\item{X}{the full feature dataset}

\item{n_rounds}{number of # boosting iterations}

\item{n_reg}{number of tested regularization parameter}
}
\value{
a data frame
}
\description{
DESCRIPTION
}
\note{
A search grid calculated based on the input data
}
