% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_data}
\alias{check_input_data}
\title{check the data for the prediction}
\usage{
check_input_data(y, X, verbose = FALSE)
}
\arguments{
\item{y}{the responding variable}

\item{X}{the feature matrix. No missing allowed}

\item{verbose}{output level: True for more logs.}
}
\value{
a idx of data records available for the prediction (no missing)
}
\description{
DESCRIPTION
}
\note{
The function checks the missing values and dimension of the
             dataset
}
