% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_foldset}
\alias{check_input_foldset}
\title{check foldset}
\usage{
check_input_foldset(foldset, y, K, verbose = FALSE)
}
\arguments{
\item{foldset}{provide a foldset to control random sampling error.}

\item{y}{the responding variable. The function assume it is non-missing}

\item{K}{integer, the number of folds in the CV}

\item{verbose}{output level: True for more logs.}
}
\value{
Checked foldset - the same structure as caret::createFolds()
}
\description{
DESCRIPTION
}
\note{
check the foldset. If there is none, create one
}
