% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_seed}
\alias{check_input_seed}
\title{check the seed of the analysis}
\usage{
check_input_seed(seed)
}
\arguments{
\item{seed}{integer or NULL, the input seed}
}
\description{
DESCRIPTION
}
\note{
Create a new seed or check an existing seed
}
