% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigMLcore.R
\docType{package}
\name{pigMLcore}
\alias{pigMLcore}
\title{pigMLcore: The core analysis module for the pigML project.}
\description{
pigMLcore: The core analysis module for the pigML project.
}
\author{
Tianfu Yang \email{tianfu@ualberta.ca}
Maintainer: Tianfu Yang \email{tianfu@ualberta.ca}
}
