% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pre_prosess_in_CV}
\alias{pre_prosess_in_CV}
\title{preprocess the feature data in CV}
\usage{
pre_prosess_in_CV(X, idx_test, n_pca)
}
\arguments{
\item{X}{the full feature dataset}

\item{idx_test}{the idx of test records in the current fold}

\item{n_pca}{number of selected top PCs. Null for no PCA}
}
\value{
A list including the transformed training and testing data
}
\description{
DESCRIPTION
}
\note{
Pre-process the feature matrix in cross-validation. if n_pca is
            given, the processing includes PCA and scaling. Otherwise it is
            simply scaling.
}
