% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_MLP.R
\name{reg_mlpL2}
\alias{reg_mlpL2}
\title{Prediction with XGBTree and evaluation through CV}
\usage{
reg_mlpL2(
  y,
  X,
  job_name = "new_job",
  K = 10,
  n_pca = NULL,
  pb = FALSE,
  foldset = NULL,
  seed = NULL,
  mlpL2_TC = trainControl(method = "cv", number = 2, verboseIter = F),
  mlpL2_TG = NULL,
  n_size = 5,
  n_lambda = 3,
  epoch = 200
)
}
\arguments{
\item{y}{the responding variable}

\item{X}{the feature matrix. No missing allowed}

\item{job_name}{A character of job ID}

\item{K}{integer, the number of folds in the CV}

\item{n_pca}{number of selected top PCs. Null for no PCA}

\item{pb}{show pgbar or nots}

\item{foldset}{provide a foldset to control random sampling error.}

\item{seed}{provide a seed to reproduce a result. NULL by default}

\item{mlpL2_TC}{trControl that pass to caret}

\item{mlpL2_TG}{tuneGrid that pass to caret}

\item{n_size}{number of size in grid search,}

\item{n_lambda}{number of lambda in grid search,}

\item{epoch}{epoch in the training}
}
\value{
a list
}
\description{
DESCRIPTION
}
\note{
The function provide a general way to evaluate the prediction
             using xgboost through CV.
}
