% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_random_forest.R
\name{reg_rf}
\alias{reg_rf}
\title{Prediction with RF and evaluation through CV}
\usage{
reg_rf(
  y,
  X,
  job_name = "new_job",
  K = 10,
  n_pca = NULL,
  verbose = FALSE,
  pb = FALSE,
  foldset = NULL,
  seed = NULL,
  rf_TC = trainControl(method = "cv", number = 5, verboseIter = F),
  rf_TG = NULL,
  n_mtry = 10,
  n_MNS = 5,
  nc = 1
)
}
\arguments{
\item{y}{the responding variable}

\item{X}{the feature matrix. No missing allowed}

\item{job_name}{A character of job ID}

\item{K}{integer, the number of folds in the CV}

\item{n_pca}{number of selected top PCs. Null for no PCA}

\item{verbose}{output log or not}

\item{pb}{show pgbar or nots}

\item{foldset}{provide a foldset to control random sampling error.}

\item{seed}{provide a seed to reproduce a result. NULL by default}

\item{rf_TC}{trControl that pass to caret}

\item{rf_TG}{tuneGrid that pass to caret}

\item{n_mtry}{number of ntry tested in the grid}

\item{n_MNS}{number of min.node.size tested in the grid}

\item{nc}{number of cpus to use in parallel- should be 1}
}
\value{
a list
}
\description{
DESCRIPTION
}
\note{
The function provide a general way to evaluate the prediction
             using RF through CV.
}
