% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_ridge_reg.R
\name{reg_rr}
\alias{reg_rr}
\title{Prediction with EN and evaluation through CV}
\usage{
reg_rr(
  y,
  X,
  job_name = "new_job",
  K = 10,
  n_pca = NULL,
  verbose = FALSE,
  pb = FALSE,
  foldset = NULL,
  seed = NULL,
  rr_TC = trainControl(method = "cv", number = 5, verboseIter = F),
  nc = 1
)
}
\arguments{
\item{y}{the responding variable}

\item{X}{the feature matrix. No missing allowed}

\item{job_name}{A character of job ID}

\item{K}{integer, the number of folds in the CV}

\item{n_pca}{number of selected top PCs. Null for no PCA}

\item{verbose}{output log or not}

\item{pb}{show pgbar or nots}

\item{foldset}{provide a foldset to control random sampling error.}

\item{seed}{provide a seed to reproduce a result. NULL by default}

\item{rr_TC}{trControl that pass to caret}

\item{nc}{number of cpus to use in parallel}
}
\value{
a list
}
\description{
DESCRIPTION
}
\note{
The function provide a general way to evaluate the prediction
             using EN through CV.
}
