% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_SVR.R
\name{reg_svrp}
\alias{reg_svrp}
\title{Prediction with SVR polynomialand evaluation through CV}
\usage{
reg_svrp(
  y,
  X,
  job_name = "new_job",
  K = 10,
  n_pca = NULL,
  pb = FALSE,
  foldset = NULL,
  seed = NULL,
  svrp_TC = trainControl(method = "cv", number = 5, verboseIter = F),
  svrp_TG = NULL,
  n_C = 10,
  n_degree = 3,
  n_Scale = 3,
  nc = 1
)
}
\arguments{
\item{y}{the responding variable}

\item{X}{the feature matrix. No missing allowed}

\item{job_name}{A character of job ID}

\item{K}{integer, the number of folds in the CV}

\item{n_pca}{number of selected top PCs. Null for no PCA}

\item{pb}{show pgbar or nots}

\item{foldset}{provide a foldset to control random sampling error.}

\item{seed}{provide a seed to reproduce a result. NULL by default}

\item{svrp_TC}{trControl that pass to caret}

\item{svrp_TG}{tuneGrid that pass to caret}

\item{n_C}{number of Cost in grid search}

\item{n_degree}{number of degree in grid search,}

\item{n_Scale}{number of Scale in grid search,}

\item{nc}{number of cpus to use in parallel}
}
\value{
a list
}
\description{
DESCRIPTION
}
\note{
The function provide a general way to evaluate the prediction
             using SVR polynomial through CV.
}
