% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_xgbtree.R
\name{reg_xgbt}
\alias{reg_xgbt}
\title{Prediction with XGBTree and evaluation through CV}
\usage{
reg_xgbt(
  y,
  X,
  job_name = "new_job",
  K = 10,
  n_pca = NULL,
  pb = FALSE,
  foldset = NULL,
  seed = NULL,
  xgbt_TC = trainControl(method = "cv", number = 5, verboseIter = F),
  xgbt_TG = NULL,
  n_rounds = 8,
  n_depth = 5,
  n_seq = 5,
  nc = 1
)
}
\arguments{
\item{y}{the responding variable}

\item{X}{the feature matrix. No missing allowed}

\item{job_name}{A character of job ID}

\item{K}{integer, the number of folds in the CV}

\item{n_pca}{number of selected top PCs. Null for no PCA}

\item{pb}{show pgbar or nots}

\item{foldset}{provide a foldset to control random sampling error.}

\item{seed}{provide a seed to reproduce a result. NULL by default}

\item{xgbt_TC}{trControl that pass to caret}

\item{xgbt_TG}{tuneGrid that pass to caret}

\item{n_rounds}{number of nrounds in grid search,}

\item{n_depth}{number of depth in grid search,}

\item{n_seq}{number of gamma/min_child_weight in grid search,}

\item{nc}{number of cpus to use in parallel}
}
\value{
a list
}
\description{
DESCRIPTION
}
\note{
The function provide a general way to evaluate the prediction
             using xgboost through CV.
}
